# [internal/stanza] Use hash instead of json encoded string