**Description:**
* Increase performance of internal/stanza converter by using a hash-based resource identifier, instead of json-based.
* Add in an "integration benchmark" and "integration test" that tests/benchmarks log ingestion through multiple components of the package (this will help with giving some confidence with future performance improvements) 

**Testing:**
Using the existing benchmark (BenchmarkConverter), I see a performance improvement:

Before:
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/stanza
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkConverter/worker_count=1-12         	       1	7253762089 ns/op	3326649368 B/op	67065109 allocs/op
BenchmarkConverter/worker_count=2-12         	       1	4751685572 ns/op	3326629080 B/op	67064759 allocs/op
BenchmarkConverter/worker_count=4-12         	       1	3505273174 ns/op	3326596144 B/op	67064529 allocs/op
BenchmarkConverter/worker_count=6-12         	       1	3250629410 ns/op	3326593848 B/op	67064483 allocs/op
BenchmarkConverter/worker_count=8-12         	       1	3254775348 ns/op	3326586600 B/op	67064483 allocs/op
PASS
```
After:
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/stanza
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkConverter/worker_count=1-12         	       1	6505129031 ns/op	3086660096 B/op	62064955 allocs/op
BenchmarkConverter/worker_count=2-12         	       1	4252483367 ns/op	3086591384 B/op	62064612 allocs/op
BenchmarkConverter/worker_count=4-12         	       1	3254180768 ns/op	3086582832 B/op	62064457 allocs/op
BenchmarkConverter/worker_count=6-12         	       1	3003439848 ns/op	3086572256 B/op	62064309 allocs/op
BenchmarkConverter/worker_count=8-12         	       1	3002254049 ns/op	3086593296 B/op	62064374 allocs/op
PASS
```